requirements.txt
streamlit==1.26.0
python-dotenv==1.0.1
openai==1.32.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.26.5
tqdm==4.66.1
PyMuPDF==1.22.5
